Summarizer/requirements.txt
streamlit
Pillow
beautifulsoup4
newspaper3k
nltk
google-cloud-storage